backend/requirements.txt
# # Django dependencies (existing)
# Django==5.0.1
# djangorestframework==3.14.0
# django-cors-headers==4.3.1
# reportlab==4.0.9

# # Voice Pipeline Dependencies
# python-dotenv==1.0.0
# groq>=0.9.0
# faster-whisper==1.0.3
# sounddevice==0.5.1
# soundfile==0.12.1
# numpy==1.26.0
# kokoro-onnx==0.2.1


# Django core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
reportlab==4.0.9

# AI
python-dotenv==1.0.0
groq>=0.9.0

# Safe numpy for Python 3.13
numpy>=2.1.0
